rolled data-related tests into a single module
